Update use of types, reorganize code
- Use more strict types when processing rows
- Rename and move types, split up processing code
- multinet.metadata -> multinet.processing, reorg
