--- multinet/metadata/__init__.py
@@ -1,1 +1,0 @@
-"""Functions and classes for handling metadata."""
